- restructure the phoenicis-scripts module - remove the Java Scripting API PhoenicisScriptEngine implementation
