+++ phoenicis-library/src/main/java/org/phoenicis/library/ShortcutRunner.java
@@ -20,7 +20,7 @@ package org.phoenicis.library;
 
 import org.graalvm.polyglot.Value;
 import org.phoenicis.library.dto.ShortcutDTO;
-import org.phoenicis.scripts.interpreter.InteractiveScriptSession;
+import org.phoenicis.scripts.session.InteractiveScriptSession;
 import org.phoenicis.scripts.interpreter.ScriptInterpreter;
 
 import java.util.List;
