+++ phoenicis-scripts/src/main/java/org/phoenicis/scripts/engine/ScriptEngineType.java
@@ -1,18 +1,14 @@
 package org.phoenicis.scripts.engine;
 
+import org.phoenicis.scripts.engine.implementation.PhoenicisScriptEngine;
+import org.phoenicis.scripts.engine.implementation.PolyglotScriptEngine;
+
 import java.util.Map;
 
 /**
  * The supported script engine types
  */
 public enum ScriptEngineType {
-    NASHORN("nashorn") {
-        @Override
-        public PhoenicisScriptEngine createScriptEngine() {
-            return new JSAScriptEngine("nashorn");
-        }
-    },
-
     GRAAL("graal.js") {
         @Override
         public PhoenicisScriptEngine createScriptEngine() {
