+++ phoenicis-scripts/src/main/java/org/phoenicis/scripts/engine/injectors/BeanInjector.java
@@ -1,7 +1,7 @@
-package org.phoenicis.scripts.engine.builtins;
+package org.phoenicis.scripts.engine.injectors;
 
 import org.phoenicis.configuration.security.Safe;
-import org.phoenicis.scripts.engine.PhoenicisScriptEngine;
+import org.phoenicis.scripts.engine.implementation.PhoenicisScriptEngine;
 import org.springframework.context.ApplicationContext;
 
 import java.lang.annotation.Annotation;
