+++ phoenicis-scripts/src/main/java/org/phoenicis/scripts/engine/injectors/EngineInjector.java
@@ -1,6 +1,6 @@
-package org.phoenicis.scripts.engine.builtins;
+package org.phoenicis.scripts.engine.injectors;
 
-import org.phoenicis.scripts.engine.PhoenicisScriptEngine;
+import org.phoenicis.scripts.engine.implementation.PhoenicisScriptEngine;
 
 /**
  * Injects some code into a Script Engine
