+++ phoenicis-scripts/src/main/java/org/phoenicis/scripts/interpreter/PhoenicisScriptInterpreter.java
@@ -16,10 +16,11 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package org.phoenicis.scripts.engine;
+package org.phoenicis.scripts.interpreter;
 
-import org.phoenicis.scripts.interpreter.InteractiveScriptSession;
-import org.phoenicis.scripts.interpreter.ScriptInterpreter;
+import org.phoenicis.scripts.session.PhoenicisInteractiveScriptSession;
+import org.phoenicis.scripts.engine.PhoenicisScriptEngineFactory;
+import org.phoenicis.scripts.session.InteractiveScriptSession;
 
 import java.util.function.Consumer;
