+++ phoenicis-scripts/src/main/resources/org/phoenicis/scripts/engine/injectors/utils.js
